Implement output group records for postintegration tasks
Previously, the output from postintegration tasks was always coalesced with the original integration task. This is now optional; it is possible to execute a postintegration task which leaves its parent integration task unaltered (in fact this is now the default), although it is not yet possible to generate any derived products from the result.

Postintegration tasks now have their output groups committed to the repository database, as for integration tasks and output tasks.