Update compiler optimization flags
The version of odeint-v2 in BOOST 1.57 seems to generate failing code when targetting AVX2 SIMD instructions. On Haswell, this means that failures occur when using -fast with the Intel compiler.

To get correct code, it seems to most aggressive SIMD instruction set we can target is AVX, ie. -mavx as a compiler flag.

This patch also enable inter-procedural optimizations, which gives a small speedup.