Preparator work for refactoring macro replacement
Macro replacement is unacceptably slow with a large number of fields - it needs to be replaced with a much less naive algorithm.

This preparatory work changes the interface so that a single input line can return a vector of output lines (rather than coalescing all output lines together with newline characters).