Fix performance regression introduced by new macro replacement implementation
Moved pre- and post- methods for index macros into the constructor and destructor of index_macro_token(). The previous method called the pre- and post-methods every time the macro was evaluated on different index assignments, which entirely misses the point of the pre- and post-methods -- they enable cachine of expensive-to-compute values! So they should only be called *once*, and the cache used to lookup the value appropriate for each index assignment.