Implement caching of GiNaC expressions
This avoids costly recomputation if the expressions are subsequently reused, which can represent a large part of the time cost in a model with many fields.

Currently caching is ineffective, becaues the symbols out of which each expression is build are different even if they have the same name. This prevents the cache from recognizing re-use. A symbol cache needs to be built to prevent this happening.