Fix inadvertent aliasing of A-, B-, C- and M-tensor results in the cache
This stemmed from use of a flattener to generate unique index numbers. This works find if the flattener always has the same dimension, but sometimes the compute_<X>_component() methods were called with a flattener of size equal to the number of fields [when computing the A-, B-, C- and M-tensors themselves], and sometimes with a flattener of size equal to twice the number of fields [when computing a u3-tensor]. This means that index values didn't map from unique ijk index tuples, and therefore some cache queries returned incorrect results.