Fix bug in runtime check for API version
Previous version tried to use model<number>::get_name() to report the name of the model which required a higher API version. However, this is invoked from a method in the virtual model<> class, in which get_name() is a purely virtual function. So it can't be used here.

Instead, we now report the UID of the failing model.