Refactor buffer handling
This refactoring simplifies the structure, which was quite complex. It is still not as simple as would be preferable, but the logic is now more explicit (eg. previous version used callbacks to ensure flushing of temporaries, which happened "magically" behind the scenes).

Buffers now have an intended capacity. The buffer will try to commit lines to disc if its size exceeds its capacity. This is intended to regulate memory usage, which can becomes very large with many-field models.