Cut down number of columns per table to be less than the number of parameters SQLite allows
Running on the quadratic10 example gave an exception with too many variables. Presumably this happened because with 10 fields and 20 phase-space coordinates, some 3-index matrices would have 20*20*20=8000 columns. Previously this would be wrapped at SQLite's default 2000 column limit, but the lower variable limit of 999 means that rows could not be inserted.