Interim commit (won't build) - refactor task stack and derived product stack in preparation for eliminating writeback
In the new structure, derived products must be associated with a task that produces the correct output. This entails introducing new concepts such as a 'derivable_task<>' from which both 'integration_task<>' and 'postintegration_task<>' inherit.

Also, 'zeta_*_task<>' classes need enhanced functionality, because they must now supply and time- and k-configuration information to derived content producers.

The next step (not addressed in this commit) is to modify 'datapipe<>' so that it can collect information from content groups associated with a postintegration product.