Allow scheduler to adjust granularity dynamically
Experience on Apollo shows that the aggregation time is about 1 minute, so with a target granularity of a minute a lot of workers finish their assignments while the master process is aggregating.

This means that the processes block until aggregation is finished (and even then, the master will only schedule one process at once), so a lot of time is wasted. By adjusting the granularity to be many times the average aggegration time the intention is to alleviate this bottleneck to some degree.