Normalize a(t) to a fixed number at horizon-crossing time for k=1 mode
Changes in normalization seem to affect the performance of the integrator. It works well with values of order e(4) or e(5). Values of order e(0) take nearly twice as long, but work. Values of order e(10) or larger seem to fail entirely.

This was causing big problems when trying to explore large squeezing ratios.