Remove separate zeta cache, which no longer has a clear rationale
Originally, zeta products were computed on-the-fly in typical workflows and the partition into 'zeta' and 'normal' caches was intended to prevent expensive-to-compute zeta products from being displaced by field-space correlation data which could be cheaply read from the database.

Due the architectural changes, zeta products are no longer computed on-the-fly so there is no real distinction between zeta products and other types. In particular they are no longer much more expensive to obtain, since everything is read from the database.

This means it's probably better to have a single larger cache which can be used for zeta products or other things as needed.