Fix filestack regression in tracking origin of lexemes
Use of std::shared_ptr<> had the unintended side-effect that only a single filestack object was used, with all lexemes pointing ot the same filestack.

At the end of processing (when the parser is acting) this filestack is empty so all origin information was lost. Instead, we should copy the filestack while constructing each lexeme.