Updating CMakeLists.txt & FindCppTransport to account for updated JsonCpp
The newer version of JsonCpp installs files to a lib64/ directory, instead of the older lib/ directory. This was causing a problem as CMake wasn't finding the files, and so erroring when compiling. This commit tells CMake to install the files to the lib64/ directory, and then tells CMake where they are when compling a model