Fixes problem with compiling with Boost >= 1.67 - CPPT-22
This fixes a problem where the function `likely` was being defined by both Boost and GINAC, and so threw errors while building CppTransport.
By changing the ordering of the include statements, this function clash has been avoided, and so CppTransport should successfully build with the latest versions of Boost.
This commit is on *my* branch 201901-ds283 to avoid having to pull in all my other changes to the task scheduler, and hence this is a repeat commit.