Fixes installation issues with newer versions of JsonCpp - CPPT-24
On certain machines, the JsonCpp libraries are installed to a lib64/ directory instead of the usual lib/ directory. This was causing FindCppTransport.cmake to fail on model building as it was not correctly searching in the possible lib64/ directory. Since it is machine dependant on the location, the lib64/ installation is marked as optional as it may not exist on some builds.