Remove the unnecessary __N argument to mass_spectrum() & sorted_mass_spectrum() - CPPT-2
It seems that in an older version of CppTransport, the e-fold number was needed to compute the mass matrix & associated eigenvalues. However, this is no longer needed, and so this commit removes this unnecessary argument. This requires a major version bump of the template files as this is a breaking change that is no longer backwards compatible.