Implement auto-detection of the location of header files - CPPT-5
This implements a new macro token $CPPT_BUILD_LOCATION which points to the direction of the /build/ directory that the user has chosen. This allows automatic linking to the correct header files & cmake files during parsing of the CMakeLists.txt file.

This also removes the $TEST macro which was used for testing macros only and shouldn't have been committed!